Migrate fitter code into interface and implementation
Move SNFitter class into its own header as the frontend algorithm.
Place all code called/exposed privately by SNFitter into a separate
translation unit.

Retain frontend fitter_library interface to help testing. Update tests
using private code to include the implementation detail header. This
is fine in a test as we are "inside" the project here.
